+++ grascii/dictionary/install.py
@@ -4,13 +4,16 @@ import argparse
 import sys
 from pathlib import Path
 from shutil import copy
+from typing import Optional
 
-from grascii.appdirs import user_data_dir
-from grascii.config import APP_NAME
+from grascii.dictionary.common import (
+    INSTALLATION_DIR,
+    DictionaryAlreadyExists,
+    get_dictionary_installed_name,
+    get_dictionary_path_name,
+)
 
 description = "Install a Grascii Dictionary"
-
-DICTIONARY_PATH = Path(user_data_dir(APP_NAME), "dictionaries")
 
 
 def build_argparser(argparser: argparse.ArgumentParser) -> None:
@@ -34,31 +37,50 @@ def build_argparser(argparser: argparse.ArgumentParser) -> None:
     )
 
 
-def install_dict(src: Path, dest: Path) -> None:
-    files = src.glob("[A-Z]")
+def install_dictionary(
+    dictionary: Path,
+    install_dir: Path,
+    name: Optional[str] = None,
+    force: bool = False,
+) -> str:
+    """Install a dictionary to an installation directory.
+
+    :param dictionary: A path to the output directory of a dictionary build.
+    :param install_dir: A path to install the dictionary to.
+    :param name: An optional name to give the installed dictionary.
+    :param force: If True, an existing dictionary can be overwritten.
+    :type dictionary: Path
+    :type install_dir: Path
+    :type name: str
+    :type force: bool
+
+    :returns: The installed name of the dictionary.
+    """
+    if name is None:
+        name = dictionary.name
+    else:
+        name = get_dictionary_path_name(name)
+    destination = install_dir / name
+    if destination.exists() and not force:
+        raise DictionaryAlreadyExists(name)
+    destination.mkdir(parents=True, exist_ok=True)
+    files = dictionary.glob("[A-Z]")
     for f in files:
-        copy(f, dest)
+        copy(f, destination)
+    return get_dictionary_installed_name(name)
 
 
 def cli_install(args: argparse.Namespace) -> None:
-    if args.name:
-        name = args.name
-    else:
-        name = args.dictionary.name
-
-    dest = DICTIONARY_PATH / name
-    if dest.exists() and not args.force:
-        if not dest.is_dir():
-            print()
-            return
-        print("A dictionary named", name, "already exists.", file=sys.stderr)
+    try:
+        name = install_dictionary(
+            args.dictionary, INSTALLATION_DIR, args.name, args.force
+        )
+    except DictionaryAlreadyExists as e:
+        print("A dictionary named", e.name, "already exists.", file=sys.stderr)
         print("Provide a different name with --name.", file=sys.stderr)
         print("If you would like to overwrite it, run with --force.", file=sys.stderr)
-        return
-
-    dest.mkdir(parents=True, exist_ok=True)
-    install_dict(args.dictionary, dest)
-    print("Successfully installed", args.dictionary, "as", name)
+    else:
+        print("Successfully installed", args.dictionary, "as", name)
 
 
 def main() -> None: