+++ grascii/dictionary/uninstall.py
@@ -2,9 +2,14 @@ from __future__ import annotations
 
 import argparse
 import sys
+from pathlib import Path
 from shutil import rmtree
 
-from grascii.dictionary.install import DICTIONARY_PATH
+from grascii.dictionary.common import (
+    INSTALLATION_DIR,
+    DictionaryNotFound,
+    get_dictionary_path_name,
+)
 
 description = "Uninstall a Grascii Dictionary"
 
@@ -18,31 +23,39 @@ def build_argparser(argparser: argparse.ArgumentParser) -> None:
     )
 
 
-def uninstall_dict(name: str, force: bool = False) -> None:
-    dictionary = DICTIONARY_PATH / name
+def uninstall_dictionary(name: str, install_dir: Path, force: bool = False) -> None:
+    """Uninstall a dictionary from an installation directory.
+
+    :param name: The name of the dictionary to uninstall.
+    :param install_dir: A path to uninstall the dictionary from.
+    :param force: If True, forces the uninstallation of a dictionary even in the
+    case of corruption.
+    :type name: Path
+    :type install_dir: Path
+    :type force: bool
+    """
+    dictionary_path = install_dir / get_dictionary_path_name(name)
+    if not dictionary_path.exists():
+        raise DictionaryNotFound(get_dictionary_path_name(name))
     if force:
-        rmtree(dictionary)
+        rmtree(dictionary_path)
     else:
-        for f in dictionary.iterdir():
+        for f in dictionary_path.iterdir():
             f.unlink()
-        dictionary.rmdir()
+        dictionary_path.rmdir()
 
 
 def cli_uninstall(args: argparse.Namespace) -> None:
-    dictionary = DICTIONARY_PATH / args.name
-    if not dictionary.exists():
-        print(args.name, "does not exist.", file=sys.stderr)
-        return
-    if not dictionary.is_dir():
-        return
-
     try:
-        uninstall_dict(args.name, args.force)
-        print("Successfully uninstalled", args.name)
-    except Exception:
+        uninstall_dictionary(args.name, INSTALLATION_DIR, force=args.force)
+    except DictionaryNotFound:
+        print(args.name, "does not exist.", file=sys.stderr)
+    except OSError:
         print("An error occurred during uninstallation.", file=sys.stderr)
         print("The dictionary may be corrupted.", file=sys.stderr)
         print("To force removal, run with --force.", file=sys.stderr)
+    else:
+        print("Successfully uninstalled", args.name)
 
 
 def main() -> None: